requirements.txt
click==8.1.7
colorama==0.4.6
dill==0.3.7
dnspython==2.3.0
drill==1.2.0
Flask==2.2.5
greenlet==3.0.1
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.21.6
pandas==1.3.5
psycopg2-binary==2.9.9
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==2.0.22
threadpoolctl==3.1.0
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
